bitmap unsigned long -> uint8_t

+++ include/bitmap.h
@@ -1,26 +1,28 @@
 #ifndef _bitmap_h_
 #define _bitmap_h_
 
+#include <stdint.h>
+
 #if defined(__cplusplus)
 extern "C" {
 #endif
 
-void bitmap_zero(unsigned long* bitmap, unsigned int nbits);
-void bitmap_fill(unsigned long* bitmap, unsigned int nbits);
+void bitmap_zero(uint8_t* bitmap, unsigned int nbits);
+void bitmap_fill(uint8_t* bitmap, unsigned int nbits);
 
-void bitmap_set(unsigned long *bitmap, unsigned int start, unsigned int len);
-void bitmap_clear(unsigned long *bitmap, unsigned int start, unsigned int len);
+void bitmap_set(uint8_t *bitmap, unsigned int start, unsigned int len);
+void bitmap_clear(uint8_t *bitmap, unsigned int start, unsigned int len);
 
-void bitmap_or(unsigned long* result, const unsigned long* src1, const unsigned long* src2, unsigned int nbits);
-void bitmap_and(unsigned long* result, const unsigned long* src1, const unsigned long* src2, unsigned int nbits);
-void bitmap_xor(unsigned long* result, const unsigned long* src1, const unsigned long* src2, unsigned int nbits);
+void bitmap_or(uint8_t* result, const uint8_t* src1, const uint8_t* src2, unsigned int nbits);
+void bitmap_and(uint8_t* result, const uint8_t* src1, const uint8_t* src2, unsigned int nbits);
+void bitmap_xor(uint8_t* result, const uint8_t* src1, const uint8_t* src2, unsigned int nbits);
 
-unsigned int bitmap_count_leading_zero(const unsigned long* bitmap, unsigned int nbits);
-unsigned int bitmap_count_next_zero(const unsigned long* bitmap, unsigned int nbits, unsigned int start);
-unsigned int bitmap_weight(const unsigned long* bitmap, unsigned int nbits);
+unsigned int bitmap_count_leading_zero(const uint8_t* bitmap, unsigned int nbits);
+unsigned int bitmap_count_next_zero(const uint8_t* bitmap, unsigned int nbits, unsigned int start);
+unsigned int bitmap_weight(const uint8_t* bitmap, unsigned int nbits);
 
 /// @return 0-not set, other-set to 1
-int bitmap_test_bit(const unsigned long* bitmap, unsigned int bits);
+int bitmap_test_bit(const uint8_t* bitmap, unsigned int bits);
 
 #if defined(__cplusplus)
 }
